+++ src/dimmer_protocol.h
@@ -8,9 +8,10 @@
 #define DIMMER_I2C_ADDRESS                  0x17
 #endif
 
-#if DIMMER_I2C_SLAVE
+// #if DIMMER_I2C_SLAVE
 
-#define DIMMER_REGISTER_START_ADDR          0x80
+#define DIMMER_REGISTER_START_ADDR          0x01
+#define DIMMER_REGISTER_SIZE                sizeof(register_mem_t)
 #define DIMMER_REGISTER_FROM_LEVEL          (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, from_level))
 #define DIMMER_REGISTER_CHANNEL             (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, channel))
 #define DIMMER_REGISTER_TO_LEVEL            (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, to_level))
@@ -19,7 +20,7 @@
 #define DIMMER_REGISTER_READ_LENGTH         (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cmd) + offsetof(register_mem_command_t, read_length))
 #define DIMMER_REGISTER_COMMAND_STATUS      (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cmd) + offsetof(register_mem_command_t, status))
 #define DIMMER_REGISTER_CHANNELS_START      (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, level))
-#define DIMMER_REGISTER_CH0_LEVEL           (DIMMER_REGISTER_CHANNELS_START)
+#define DIMMER_REGISTER_CH0_LEVEL           (DIMMER_REGISTER_CHANNELS_START + sizeof(register_mem_t().level[0]) * 0)
 #define DIMMER_REGISTER_CH1_LEVEL           (DIMMER_REGISTER_CHANNELS_START + sizeof(register_mem_t().level[0]) * 1)
 #define DIMMER_REGISTER_CH2_LEVEL           (DIMMER_REGISTER_CHANNELS_START + sizeof(register_mem_t().level[0]) * 2)
 #define DIMMER_REGISTER_CH3_LEVEL           (DIMMER_REGISTER_CHANNELS_START + sizeof(register_mem_t().level[0]) * 3)
@@ -30,67 +31,82 @@
 #define DIMMER_REGISTER_CHANNELS_END        (DIMMER_REGISTER_CHANNELS_START + sizeof(register_mem_t().level))
 #define DIMMER_REGISTER_TEMP                (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, temp))
 #define DIMMER_REGISTER_VCC                 (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, vcc))
+#define DIMMER_REGISTER_CONFIG_OFS          (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg))
+#define DIMMER_REGISTER_CONFIG_SZ           sizeof(register_mem_cfg_t)
 #define DIMMER_REGISTER_OPTIONS             (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, options))
 #define DIMMER_REGISTER_MAX_TEMP            (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, max_temp))
 #define DIMMER_REGISTER_FADE_IN_TIME        (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, fade_in_time))
 #define DIMMER_REGISTER_TEMP_CHECK_INT      (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, temp_check_interval))
-#define DIMMER_REGISTER_LC_FACTOR           (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, linear_correction_factor))
+// #define DIMMER_REGISTER_LC_FACTOR           (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, linear_correction_factor))
 #define DIMMER_REGISTER_ZC_DELAY_TICKS      (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, zero_crossing_delay_ticks))
 #define DIMMER_REGISTER_MIN_ON_TIME_TICKS   (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, minimum_on_time_ticks))
 #define DIMMER_REGISTER_ADJ_HALFWAVE_TICKS  (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, adjust_halfwave_time_ticks))
 #define DIMMER_REGISTER_INT_1_1V_REF        (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, internal_1_1v_ref))
 #define DIMMER_REGISTER_INT_TEMP_OFS        (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, int_temp_offset))
 #define DIMMER_REGISTER_METRICS_INT         (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, report_metrics_max_interval))
+#define DIMMER_REGISTER_ERRORS_FREQ_LOW     (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, errors) + offsetof(register_mem_errors_t, frequency_low))
+#define DIMMER_REGISTER_ERRORS_FREQ_HIGH    (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, errors) + offsetof(register_mem_errors_t, frequency_high))
+#define DIMMER_REGISTER_ERRORS_ZC_MISFIRE   (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, errors) + offsetof(register_mem_errors_t, zc_misfire))
+#if DIMMER_CUBIC_INTERPOLATION
+#define DIMMER_REGISTER_CH0_CUBIC_INT       (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, cubic_int) + sizeof(register_mem_cubic_int_t) * 0)
+#define DIMMER_REGISTER_CH1_CUBIC_INT       (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, cubic_int) + sizeof(register_mem_cubic_int_t) * 1)
+#define DIMMER_REGISTER_CH2_CUBIC_INT       (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, cubic_int) + sizeof(register_mem_cubic_int_t) * 2)
+#define DIMMER_REGISTER_CH3_CUBIC_INT       (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, cubic_int) + sizeof(register_mem_cubic_int_t) * 3)
+#define DIMMER_REGISTER_CH4_CUBIC_INT       (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, cubic_int) + sizeof(register_mem_cubic_int_t) * 4)
+#define DIMMER_REGISTER_CH5_CUBIC_INT       (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, cubic_int) + sizeof(register_mem_cubic_int_t) * 5)
+#define DIMMER_REGISTER_CH6_CUBIC_INT       (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, cubic_int) + sizeof(register_mem_cubic_int_t) * 6)
+#define DIMMER_REGISTER_CH7_CUBIC_INT       (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, cubic_int) + sizeof(register_mem_cubic_int_t) * 7)
+#endif
 #define DIMMER_REGISTER_VERSION             (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, version))
 #define DIMMER_REGISTER_ADDRESS             (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, address))
 #define DIMMER_REGISTER_END_ADDR            (DIMMER_REGISTER_START_ADDR + sizeof(register_mem_t))
 
-#else
+// #else
 
-#define DIMMER_REGISTER_FROM_LEVEL          0x80
-#define DIMMER_REGISTER_CHANNEL             0x82
-#define DIMMER_REGISTER_TO_LEVEL            0x83
-#define DIMMER_REGISTER_TIME                0x85
-#define DIMMER_REGISTER_COMMAND             0x89
-#define DIMMER_REGISTER_READ_LENGTH         0x8A
-#define DIMMER_REGISTER_COMMAND_STATUS      0x8B
-#define DIMMER_REGISTER_CH0_LEVEL           0x8C
-#define DIMMER_REGISTER_CH1_LEVEL           0x8E
-#define DIMMER_REGISTER_CH2_LEVEL           0x90
-#define DIMMER_REGISTER_CH3_LEVEL           0x92
-#define DIMMER_REGISTER_CH4_LEVEL           0x94
-#define DIMMER_REGISTER_CH5_LEVEL           0x96
-#define DIMMER_REGISTER_CH6_LEVEL           0x98
-#define DIMMER_REGISTER_CH7_LEVEL           0x9A
-#define DIMMER_REGISTER_TEMP                0x9C
-#define DIMMER_REGISTER_VCC                 0xA0
-#define DIMMER_REGISTER_OPTIONS             0xA2
-#define DIMMER_REGISTER_MAX_TEMP            0xA3
-#define DIMMER_REGISTER_FADE_IN_TIME        0xA4
-#define DIMMER_REGISTER_TEMP_CHECK_INT      0xA8
-#define DIMMER_REGISTER_LC_FACTOR           0xA9
-#define DIMMER_REGISTER_ZC_DELAY_TICKS      0xAD
-#define DIMMER_REGISTER_MIN_ON_TIME_TICKS   0xAE
-#define DIMMER_REGISTER_ADJ_HALFWAVE_TICKS  0xB0
-#define DIMMER_REGISTER_INT_1_1V_REF        0xB2
-#define DIMMER_REGISTER_INT_TEMP_OFS        0xB6
-#define DIMMER_REGISTER_NTC_TEMP_OFS        0xB7
-#define DIMMER_REGISTER_METRICS_INT         0xB8
-#define DIMMER_REGISTER_VERSION             0xB9
-#define DIMMER_REGISTER_ERR_FREQ_LOW        0xBB
-#define DIMMER_REGISTER_ERR_FREQ_HIGH       0xBC
-#define DIMMER_REGISTER_ERR_ZC_MISFIRE      0xBD
-#define DIMMER_REGISTER_ADDRESS             0xBE
-#define DIMMER_REGISTER_END_ADDR            0xBF
+// #define DIMMER_REGISTER_FROM_LEVEL          0x80
+// #define DIMMER_REGISTER_CHANNEL             0x82
+// #define DIMMER_REGISTER_TO_LEVEL            0x83
+// #define DIMMER_REGISTER_TIME                0x85
+// #define DIMMER_REGISTER_COMMAND             0x89
+// #define DIMMER_REGISTER_READ_LENGTH         0x8A
+// #define DIMMER_REGISTER_COMMAND_STATUS      0x8B
+// #define DIMMER_REGISTER_CH0_LEVEL           0x8C
+// #define DIMMER_REGISTER_CH1_LEVEL           0x8E
+// #define DIMMER_REGISTER_CH2_LEVEL           0x90
+// #define DIMMER_REGISTER_CH3_LEVEL           0x92
+// #define DIMMER_REGISTER_CH4_LEVEL           0x94
+// #define DIMMER_REGISTER_CH5_LEVEL           0x96
+// #define DIMMER_REGISTER_CH6_LEVEL           0x98
+// #define DIMMER_REGISTER_CH7_LEVEL           0x9A
+// #define DIMMER_REGISTER_TEMP                0x9C
+// #define DIMMER_REGISTER_VCC                 0xA0
+// #define DIMMER_REGISTER_OPTIONS             0xA2
+// #define DIMMER_REGISTER_MAX_TEMP            0xA3
+// #define DIMMER_REGISTER_FADE_IN_TIME        0xA4
+// #define DIMMER_REGISTER_TEMP_CHECK_INT      0xA8
+// #define DIMMER_REGISTER_LC_FACTOR           0xA9
+// #define DIMMER_REGISTER_ZC_DELAY_TICKS      0xAD
+// #define DIMMER_REGISTER_MIN_ON_TIME_TICKS   0xAE
+// #define DIMMER_REGISTER_ADJ_HALFWAVE_TICKS  0xB0
+// #define DIMMER_REGISTER_INT_1_1V_REF        0xB2
+// #define DIMMER_REGISTER_INT_TEMP_OFS        0xB6
+// #define DIMMER_REGISTER_NTC_TEMP_OFS        0xB7
+// #define DIMMER_REGISTER_METRICS_INT         0xB8
+// #define DIMMER_REGISTER_VERSION             0xB9
+// #define DIMMER_REGISTER_ERR_FREQ_LOW        0xBB
+// #define DIMMER_REGISTER_ERR_FREQ_HIGH       0xBC
+// #define DIMMER_REGISTER_ERR_ZC_MISFIRE      0xBD
+// #define DIMMER_REGISTER_ADDRESS             0xBE
+// #define DIMMER_REGISTER_END_ADDR            0xBF
 
-#endif
+// #endif
 
 // first byte in master mode indicates the data that was sent
-#define DIMMER_METRICS_REPORT               0xf0
-#define DIMMER_TEMPERATURE_ALERT            0xf1
-#define DIMMER_FADING_COMPLETE              0xf2
-#define DIMMER_EEPROM_WRITTEN               0xf3
-#define DIMMER_FREQUENCY_WARNING            0xf4
+#define DIMMER_RESPONSE_METRICS_REPORT       0xf0
+#define DIMMER_RESPONSE_TEMPERATURE_ALERT    0xf1
+#define DIMMER_RESPONSE_FADING_COMPLETE      0xf2
+#define DIMMER_RESPONSE_EEPROM_WRITTEN       0xf3
+#define DIMMER_RESPONSE_FREQUENCY_WARNING    0xf4
 
 // DIMMER_REGISTER_COMMAND
 #define DIMMER_COMMAND_SET_LEVEL            0x10
@@ -106,10 +122,15 @@
 #define DIMMER_COMMAND_PRINT_INFO           0x53
 #define DIMMER_COMMAND_FORCE_TEMP_CHECK     0x54
 #define DIMMER_COMMAND_PRINT_METRICS        0x55
+#define DIMMER_COMMAND_PRINT_CUBIC_INT      0x56
 #define DIMMER_COMMAND_ZC_TIMINGS_OUTPUT    0x60
+#define DIMMER_COMMAND_ZC_DECREASE          0x82
+#define DIMMER_COMMAND_ZC_INCREASE          0x83
+#define DIMMER_COMMAND_SET_ZC               0x92
+#define DIMMER_COMMAND_FORCE_WRITE_EEPROM   0x93
+#define DIMMER_COMMAND_SIMULATE_ZC          0xe0
 #if DEBUG
 #define DIMMER_COMMAND_DUMP_MEM             0xee
-#define DIMMER_COMMAND_DUMP_MACROS          0xef
 #endif
 
 // DIMMER_REGISTER_COMMAND_STATUS
@@ -118,8 +139,11 @@
 
 // DIMMER_REGISTER_OPTIONS
 #define DIMMER_OPTIONS_RESTORE_LEVEL        0x01
-#define DIMMER_OPTIONS_REPORT_TEMP          0x02
+#define DIMMER_OPTIONS_REPORT_METRICS       0x02
 #define DIMMER_OPTIONS_TEMP_ALERT_TRIGGERED 0x04
+#define DIMMER_OPTIONS_FREQ_LOW_ALERT       0x08
+#define DIMMER_OPTIONS_FREQ_HIGH_ALERT      0x10
+#define DIMMER_OPTIONS_CUBIC_INT            0x20
 
 // DIMMER_COMMAND_READ_TIMINGS
 // All timings are float and read from DIMMER_REGISTER_TEMP