+++ src/i2c_slave.h
@@ -7,6 +7,7 @@
 
 #include "main.h"
 #include "dimmer_protocol.h"
+#include "helpers.h"
 
 #if SERIAL_I2C_BRDIGE
 
@@ -20,14 +21,24 @@
 
 #include "dimmer_reg_mem.h"
 
-typedef struct __attribute__packed__ {
-    uint32_t eeprom_cycle;
-    uint16_t crc16;
-    register_mem_cfg_t cfg;
-    int16_t level[DIMMER_CHANNELS];
-} config_t;
+class config_t {
+public:
+    config_t();
 
-void dump_macros();
+    void clear();
+    void read(size_t position);
+    size_t write(size_t position);
+    bool compare(size_t position);
+    size_t size() const;
+
+    struct __attribute__packed__ {
+        uint32_t eeprom_cycle;
+        uint16_t crc16;
+        int16_t level[DIMMER_CHANNELS];
+    };
+    register_mem_cfg_t &cfg;
+};
+
 void dimmer_i2c_slave_setup();
 void dimmer_i2c_on_receive(int length,  Stream *in);
 void dimmer_i2c_on_request(Stream *out);